third_party/WebKit/LayoutTests/http/tests/htmlimports/import-script-nosniff.html

<!DOCTYPE html>
<html>
<head>
<link id="target" rel="import" href="resources/having-sniffing-script.html">
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<pre id="result"></pre>
<script>
shouldBe("document.sniffing", "undefined");
</script>
</body>
</html>
